Changed nhsym to int32_t. Changed ospecial callback function type to uint64_t to match ospecial type.

+++ include/layer.h
@@ -239,7 +239,7 @@ struct item_score_count_result
 /*
  * type nhsym: loadable symbols go into this type
  */
-typedef int64_t nhsym;
+typedef int32_t nhsym;
 
 
 #endif /* LAYER_H */
