+++ win/win32/xpl/GnollHackX/GnollHackX/Pages/Game/GamePage.xaml.cs
@@ -8039,7 +8039,7 @@ namespace GnollHackX.Pages.Game
                                                 canvas.DrawRect(winRect, textPaint.Paint);
                                                 textPaint.Color = SKColors.Black;
                                             }
-                                            else if ((_mapData[mapx, mapy].Special & (uint)MapSpecial.Inverse) != 0)
+                                            else if ((_mapData[mapx, mapy].Special & (UInt64)MapSpecial.Inverse) != 0)
                                             {
                                                 textPaint.Style = SKPaintStyle.Fill;
                                                 SKRect winRect = new SKRect(tx, ty + textPaint.FontMetrics.Ascent, tx + width, ty + textPaint.FontMetrics.Ascent + height);
@@ -8050,7 +8050,7 @@ namespace GnollHackX.Pages.Game
                                             //canvas.DrawText(str, tx, ty, textPaint);
                                             textPaint.DrawTextOnCanvas(canvas, str, tx, ty);
 
-                                            if ((_mapData[mapx, mapy].Special & (uint)MapSpecial.Underline) != 0)
+                                            if ((_mapData[mapx, mapy].Special & (UInt64)MapSpecial.Underline) != 0)
                                             {
                                                 //canvas.DrawText("_", tx, ty, textPaint);
                                                 textPaint.DrawTextOnCanvas(canvas, "_", tx, ty);
